requirements.txt
numpy==1.24.0
torch==1.10.1
opencv-python==4.6.0.66
Pillow==9.3.0
matplotlib==3.6.3 
tqdm==4.65.0